requirements.txt
Django>=5.0,<6.0
python-dotenv>=1.0
requests>=2.31
gunicorn>=21.2
psycopg2-binary>=2.9
